+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix star_oddi parser time variable to output a datetime dataarray.
 - Fix typo in onset parser.
+- Avoid reimporting parser if already imported in read.file with parser
+defined via string expression.
 
 ## `v0.4.0` - 2024-05-04